+++ packages/integral_isolates/test/back_pressure_strategy_test.dart
@@ -18,27 +18,27 @@ void main() {
     List iterable() => [1, 'test', 2, 3, 4, 5].toList();
 
     Future<List> runIsolate(BackpressureStrategy strategy) async {
-      final isolated = StatefulIsolate(
+      final isolate = StatefulIsolate(
         backpressureStrategy: strategy,
         autoInit: false,
       );
-      await isolated.init();
+      await isolate.init();
 
       final responses = [];
 
       int delayMultiplier = 0;
-      Future temp(dynamic input) async {
+      Future runCompute(dynamic input) async {
         await Future.delayed(Duration(milliseconds: 50 * delayMultiplier++));
         if (input is String) {
           try {
-            final value = await isolated.isolate(_testFunctionStr, input);
+            final value = await isolate.compute(_testFunctionStr, input);
             responses.add(value);
           } catch (e) {
             // Noop
           }
         } else if (input is int) {
           try {
-            final value = await isolated.isolate(_testFunction, input);
+            final value = await isolate.compute(_testFunction, input);
             responses.add(value);
           } catch (e) {
             // Noop
@@ -47,10 +47,10 @@ void main() {
       }
 
       await Future.wait([
-        for (final number in iterable()) temp(number),
+        for (final number in iterable()) runCompute(number),
       ]);
 
-      isolated.dispose();
+      isolate.dispose();
 
       return responses;
     }
@@ -62,9 +62,9 @@ void main() {
       await Future.wait([
         for (final input in iterable())
           if (input is String)
-            isolate.isolate(_testFunctionStr, input).then(responses.add)
+            isolate.compute(_testFunctionStr, input).then(responses.add)
           else if (input is int)
-            isolate.isolate(_testFunction, input).then(responses.add),
+            isolate.compute(_testFunction, input).then(responses.add),
       ]);
 
       isolate.dispose();
@@ -73,28 +73,34 @@ void main() {
     });
 
     test('No backpressure strategy', () async {
-      final responses = await runIsolate(NoBackPressureStrategy());
-      expect(responses, [1, 'prefix: test', 2, 3, 4, 5]);
+      await expectLater(
+        runIsolate(NoBackPressureStrategy()),
+        completion([1, 'prefix: test', 2, 3, 4, 5]),
+      );
     });
 
     test('Discard new backpressure strategy', () async {
-      final responses = await runIsolate(DiscardNewBackPressureStrategy());
-      expect(responses, [1, 'prefix: test']);
+      await expectLater(
+        runIsolate(DiscardNewBackPressureStrategy()),
+        completion([1, 'prefix: test']),
+      );
     });
 
     test('Replace backpressure strategy', () async {
-      final responses = await runIsolate(ReplaceBackpressureStrategy());
-      expect(responses, [1, 5]);
+      await expectLater(
+        runIsolate(ReplaceBackpressureStrategy()),
+        completion([1, 5]),
+      );
     });
 
     test('Combine backpressure strategy', () async {
-      final responses = await runIsolate(
+      final future = runIsolate(
         CombineBackPressureStrategy((oldData, newData) {
           if (oldData is num && newData is num) return oldData + newData;
           return newData;
         }),
       );
-      expect(responses, [1, 14]);
+      expect(future, completion([1, 14]));
     });
   });
 }