+++ packages/integral_isolates/test/backpressure_tailored_isolated_test.dart
@@ -15,11 +15,11 @@ void main() {
     Future<List<int>> runIsolate(
       BackpressureStrategy<int, int> strategy,
     ) async {
-      final isolated = TailoredStatefulIsolate<int, int>(
+      final isolate = TailoredStatefulIsolate<int, int>(
         backpressureStrategy: strategy,
         autoInit: false,
       );
-      await isolated.init();
+      await isolate.init();
 
       final responses = <int>[];
 
@@ -27,7 +27,7 @@ void main() {
         await Future.delayed(Duration(milliseconds: 50 * number));
 
         try {
-          final value = await isolated.isolate(_testFunction, number);
+          final value = await isolate.compute(_testFunction, number);
           responses.add(value);
         } catch (e) {
           // Noop
@@ -38,7 +38,7 @@ void main() {
         for (final number in iterable()) temp(number),
       ]);
 
-      isolated.dispose();
+      isolate.dispose();
 
       return responses;
     }
@@ -49,7 +49,7 @@ void main() {
 
       await Future.wait([
         for (final int in iterable())
-          isolate.isolate(_testFunction, int).then(responses.add),
+          isolate.compute(_testFunction, int).then(responses.add),
       ]);
 
       isolate.dispose();
@@ -58,18 +58,24 @@ void main() {
     });
 
     test('No backpressure strategy', () async {
-      final responses = await runIsolate(NoBackPressureStrategy());
-      expect(responses, [1, 2, 3, 4, 5]);
+      expect(
+        runIsolate(NoBackPressureStrategy()),
+        completion([1, 2, 3, 4, 5]),
+      );
     });
 
     test('Discard new backpressure strategy', () async {
-      final responses = await runIsolate(DiscardNewBackPressureStrategy());
-      expect(responses, [1, 2]);
+      expect(
+        runIsolate(DiscardNewBackPressureStrategy()),
+        completion([1, 2]),
+      );
     });
 
     test('Replace backpressure strategy', () async {
-      final responses = await runIsolate(ReplaceBackpressureStrategy());
-      expect(responses, [1, 5]);
+      expect(
+        runIsolate(ReplaceBackpressureStrategy()),
+        completion([1, 5]),
+      );
     });
 
     test('Combine backpressure strategy', () async {
@@ -78,14 +84,17 @@ void main() {
           (oldData, newData) => oldData + newData,
         ),
       );
-      isolate.isolate((message) => message + 2, 2);
+      isolate.compute((message) => message + 2, 2);
 
-      final responses = await runIsolate(
+      final future = runIsolate(
         CombineBackPressureStrategy<int, int>((oldData, newData) {
           return oldData + newData;
         }),
       );
-      expect(responses, [1, 14]);
+       expect(
+        future,
+        completion([1, 14]),
+      );
     });
   });
 }